+++ appsettings.json
@@ -59,7 +59,7 @@
       "FieldMappingTool": {
         "Enabled": false,
         "FieldMaps": [],
-        "FieldMapDefaults" :  {
+        "FieldMapDefaults": {
           "ApplyTo": [ "*" ]
         }
       },
@@ -427,6 +427,39 @@
         "OnlyListUsersInWorkItems": true,
         "SourceName": "Source",
         "TargetName": "Target"
+      }
+    },
+    "Infrastructure": {
+      "ClassNameChangeMappings": {
+        "WorkItemMigration": "TfsWorkItemMigrationProcessor",
+        "WorkItemMigrationContext": "TfsWorkItemMigrationProcessor",
+        "TfsTeamProjectConfig": "TfsTeamProjectEndpoint",
+        "WorkItemGitRepoMappingTool": "TfsGitRepositoryTool",
+        "WorkItemFieldMappingTool": "FieldMappingTool",
+        "CreateTeamFolders": "TfsCreateTeamFoldersProcessor",
+        "ExportProfilePicture": "TfsExportProfilePictureFromADProcessor",
+        "ExportProfilePictureFromAD": "TfsExportProfilePictureFromADProcessor",
+        "ImportProfilePicture": "TfsImportProfilePictureProcessor",
+        "ImportProfilePictureFromAD": "TfsImportProfilePictureProcessor",
+        "ExportUsersForMapping": "TfsExportUsersForMappingProcessor",
+        "ExportUsersForMappingContext": "TfsExportUsersForMappingProcessor",
+        "ExportTeamListProcessor": "TfsExportTeamListProcessor",
+        "ExportTeamList": "TfsExportTeamListProcessor",
+        "ExportUsersForMappingProcessor": "TfsExportUsersForMappingProcessor",
+        "TestConfigurationsMigrationProcessor": "TfsTestConfigurationsMigrationProcessor",
+        "TestConfigurationsMigrationContext": "TfsTestConfigurationsMigrationProcessor",
+        "TestPlansAndSuitesMigrationProcessor": "TfsTestPlansAndSuitesMigrationProcessor",
+        "TestPlansAndSuitesMigrationContext": "TfsTestPlansAndSuitesMigrationProcessor",
+        "TestVariablesMigrationProcessor": "TfsTestVariablesMigrationProcessor",
+        "TestVariablesMigrationContext": "TfsTestVariablesMigrationProcessor",
+        "WorkItemBulkEditProcessor": "TfsWorkItemBulkEditProcessor",
+        "WorkItemUpdate": "TfsWorkItemBulkEditProcessor",
+        "WorkItemDelete": "TfsWorkItemDeleteProcessor",
+        "WorkItemDeleteProcessor": "TfsWorkItemDeleteProcessor",
+        "WorkItemPostProcessingContext": "TfsWorkItemOverwriteProcessor",
+        "WorkItemPostProcessingProcessor": "TfsWorkItemOverwriteProcessor",
+        "WorkItemUpdateAreasAsTagsContext": "TfsWorkItemOverwriteProcessor",
+        "WorkItemUpdateAreasAsTagsProcessor": "TfsWorkItemOverwriteProcessor"
       }
     }
   }
