+++ src/MigrationTools.Host/Commands/InitMigrationCommandSettings.cs
@@ -21,7 +21,7 @@ namespace MigrationTools.Host.Commands
     {
         Reference = 0,
         WorkItemTracking = 1,
-        Basic = 4,
+        Basic = 3,
         PipelineProcessor = 5
     }
 }
