+++ src/MigrationTools/MigrationTools.csproj
@@ -15,6 +15,10 @@
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
   </PropertyGroup>
 
+	<ItemGroup>
+		<NuGetAuditSuppress Include="*" />
+	</ItemGroup>
+
 	<PropertyGroup>
 		<GitVersion>false</GitVersion>
 	</PropertyGroup>
