+++ src/MigrationTools/Options/OptionsConfigurationBuilder.cs
@@ -9,8 +9,8 @@ namespace MigrationTools.Options
 {
 
 
-public static class OptionsConfigurationBuilder
-{
+public static class OptionsConfigurationCompiler
+    {
     public static IOptions LoadConfiguration(string filePath, IOptions options, bool isCollection = false)
     {
         var optionsConfig = options.ConfigurationMetadata;
